third_party/WebKit/LayoutTests/fast/dom/custom/invalid-type-extension-crash.svg

<html xmlns="http://www.w3.org/1999/xhtml"><x-test is="x-a" id="t"></x-test>
<script src="../../../resources/js-test.js"></script>
<script><![CDATA[
description('Accessing an XML element that looks like a Custom Element ' +
            'should not crash.');
shouldBeTrue('document.getElementById("t") instanceof Element');

var successfullyParsed = true;
]]></script>
</html>
